feat(theme): update dark-light theme scripts

+++ .config/niri/scripts/toggle-sunset.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 
+DMS_PID=$(pgrep dms)
+if [[ -n "$DMS_PID" ]]; then
+    dms ipc call night toggle
+    exit 0
+fi
+
 PID=$(pgrep wlsunset)
 if [[ -n "$PID" ]]; then
   kill $PID
@@ -7,3 +13,4 @@ if [[ -n "$PID" ]]; then
 fi
 
 wlsunset -t 3000 &
+exit 0
